fix docs
